Move classes from model/bpmn to model/bpmn/internal
