+++ src/model/bpmn/internal/Bounds.ts
