+++ src/model/bpmn/internal/Label.ts
