+++ src/model/bpmn/internal/edge/Edge.ts
