+++ src/model/bpmn/internal/edge/Flow.ts
