+++ src/model/bpmn/internal/edge/FlowKind.ts
