+++ src/model/bpmn/internal/edge/MessageVisibleKind.ts
