+++ src/model/bpmn/internal/edge/SequenceFlowKind.ts
