+++ src/model/bpmn/internal/edge/Waypoint.ts
