+++ src/model/bpmn/internal/shape/Shape.ts
