+++ src/model/bpmn/internal/shape/ShapeBpmnElement.ts
