+++ src/model/bpmn/internal/shape/ShapeBpmnElementKind.ts
