+++ src/model/bpmn/internal/shape/ShapeBpmnEventKind.ts
