+++ src/model/bpmn/internal/shape/ShapeBpmnMarkerKind.ts
