+++ src/model/bpmn/internal/shape/ShapeBpmnSubProcessKind.ts
