+++ src/model/bpmn/internal/shape/ShapeUtil.ts
