+++ src/model/bpmn/internal/shape/index.ts
