+++ test/unit/component/parser/json/BpmnJsonParser.messageFlow.test.ts
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 import { parseJsonAndExpectOnlyEdges, verifyEdge } from './JsonTestUtils';
-import Waypoint from '../../../../../src/model/bpmn/edge/Waypoint';
-import { MessageVisibleKind } from '../../../../../src/model/bpmn/edge/MessageVisibleKind';
+import Waypoint from '../../../../../src/model/bpmn/internal/edge/Waypoint';
+import { MessageVisibleKind } from '../../../../../src/model/bpmn/internal/edge/MessageVisibleKind';
 import * as bpmndi from '../../../../../src/component/parser/xml/bpmn-json-model/BPMNDI';
 
 describe('parse bpmn as json for message flow', () => {
