+++ test/unit/component/parser/json/BpmnJsonParser.sequenceFlow.conditional.test.ts
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 import { parseJsonAndExpectOnlyEdgesAndFlowNodes, verifyEdge } from './JsonTestUtils';
-import { SequenceFlowKind } from '../../../../../src/model/bpmn/edge/SequenceFlowKind';
+import { SequenceFlowKind } from '../../../../../src/model/bpmn/internal/edge/SequenceFlowKind';
 import each from 'jest-each';
 import { TProcess } from '../../../../../src/component/parser/xml/bpmn-json-model/baseElement/rootElement/rootElement';
 import { TSequenceFlow } from '../../../../../src/component/parser/xml/bpmn-json-model/baseElement/flowElement';
-import Waypoint from '../../../../../src/model/bpmn/edge/Waypoint';
+import Waypoint from '../../../../../src/model/bpmn/internal/edge/Waypoint';
 
 describe('parse bpmn as json for conditional sequence flow', () => {
   each([
