+++ test/unit/component/parser/json/BpmnJsonParser.sequenceFlow.default.test.ts
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 import { parseJsonAndExpectOnlyEdgesAndFlowNodes, verifyEdge } from './JsonTestUtils';
-import { SequenceFlowKind } from '../../../../../src/model/bpmn/edge/SequenceFlowKind';
+import { SequenceFlowKind } from '../../../../../src/model/bpmn/internal/edge/SequenceFlowKind';
 import each from 'jest-each';
-import Waypoint from '../../../../../src/model/bpmn/edge/Waypoint';
+import Waypoint from '../../../../../src/model/bpmn/internal/edge/Waypoint';
 import { TProcess } from '../../../../../src/component/parser/xml/bpmn-json-model/baseElement/rootElement/rootElement';
 
 describe('parse bpmn as json for default sequence flow', () => {
