+++ test/unit/component/parser/json/JsonTestUtils.ts
@@ -13,18 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { ShapeBpmnElementKind, ShapeBpmnCallActivityKind, ShapeBpmnMarkerKind, ShapeBpmnSubProcessKind, ShapeBpmnEventKind } from '../../../../../src/model/bpmn/shape';
-import Shape from '../../../../../src/model/bpmn/shape/Shape';
+import { ShapeBpmnElementKind, ShapeBpmnCallActivityKind, ShapeBpmnMarkerKind, ShapeBpmnSubProcessKind, ShapeBpmnEventKind } from '../../../../../src/model/bpmn/internal/shape';
+import Shape from '../../../../../src/model/bpmn/internal/shape/Shape';
 import { defaultBpmnJsonParser } from '../../../../../src/component/parser/json/BpmnJsonParser';
-import Edge from '../../../../../src/model/bpmn/edge/Edge';
-import BpmnModel from '../../../../../src/model/bpmn/BpmnModel';
-import Waypoint from '../../../../../src/model/bpmn/edge/Waypoint';
-import { ShapeBpmnActivity, ShapeBpmnCallActivity, ShapeBpmnEvent, ShapeBpmnSubProcess } from '../../../../../src/model/bpmn/shape/ShapeBpmnElement';
-import { SequenceFlowKind } from '../../../../../src/model/bpmn/edge/SequenceFlowKind';
-import Label from '../../../../../src/model/bpmn/Label';
-import { SequenceFlow } from '../../../../../src/model/bpmn/edge/Flow';
-import { FlowKind } from '../../../../../src/model/bpmn/edge/FlowKind';
-import { MessageVisibleKind } from '../../../../../src/model/bpmn/edge/MessageVisibleKind';
+import Edge from '../../../../../src/model/bpmn/internal/edge/Edge';
+import BpmnModel from '../../../../../src/model/bpmn/internal/BpmnModel';
+import Waypoint from '../../../../../src/model/bpmn/internal/edge/Waypoint';
+import { ShapeBpmnActivity, ShapeBpmnCallActivity, ShapeBpmnEvent, ShapeBpmnSubProcess } from '../../../../../src/model/bpmn/internal/shape/ShapeBpmnElement';
+import { SequenceFlowKind } from '../../../../../src/model/bpmn/internal/edge/SequenceFlowKind';
+import Label from '../../../../../src/model/bpmn/internal/Label';
+import { SequenceFlow } from '../../../../../src/model/bpmn/internal/edge/Flow';
+import { FlowKind } from '../../../../../src/model/bpmn/internal/edge/FlowKind';
+import { MessageVisibleKind } from '../../../../../src/model/bpmn/internal/edge/MessageVisibleKind';
 import { BpmnJsonModel } from '../../../../../src/component/parser/xml/bpmn-json-model/BPMN20';
 
 export interface ExpectedShape {
