+++ test/unit/model/bpmn/shape/ShapeUtil.test.ts
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-import ShapeUtil from '../../../../../src/model/bpmn/shape/ShapeUtil';
-import { ShapeBpmnElementKind } from '../../../../../src/model/bpmn/shape';
+import ShapeUtil from '../../../../../src/model/bpmn/internal/shape/ShapeUtil';
+import { ShapeBpmnElementKind } from '../../../../../src/model/bpmn/internal/shape';
 
 describe('ShapeUtil', () => {
   it('top level bpmn event kinds', () => {
